Fix the namespace for the moved event classes
